+++ database/migrations/2026_02_20_235305_create_deductions_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('deduction');
             $table->double('rate');
             $table->text('description')->nullable();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
             $table->foreign('user_id')
                 ->references('id')->on('users')
